added home page and dynamically generated player data page
