+++ .gitignore
@@ -10,7 +10,7 @@ __pycache__/
 *.pyo
 venv/
 env/
-etl_service/.env
+.env
 *.log
 *.env
