+++ etl_service/.dockerignore
@@ -4,7 +4,7 @@ __pycache__/
 *.pyd
 *.log
 ../.venv/
-.env
+../.env
 .idea/
 *.iml
 *.iws
