+++ etl_service/config.yml
@@ -4,27 +4,27 @@ country_codes:
   - BR
   - GB
   - DE
-#  - FR
-#  - IT
-#  - ES
-#  - JP
-#  - KR
-#  - CN
-#  - AU
-#  - IN
-#  - TR
-#  - RU
-#  - SE
-#  - NO
-#  - FI
-#  - DK
-#  - MX
-#  - AR
-#  - CL
-#  - ZA
-#  - SG
-#  - MY
-#  - PH
-#  - ID
-#  - TH
+  - FR
+  - IT
+  - ES
+  - JP
+  - KR
+  - CN
+  - AU
+  - IN
+  - TR
+  - RU
+  - SE
+  - NO
+  - FI
+  - DK
+  - MX
+  - AR
+  - CL
+  - ZA
+  - SG
+  - MY
+  - PH
+  - ID
+  - TH
   - global
