+++ web_ui_service/templates/index.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Choose a Country</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}">
+</head>
+<body>
+
+{% if countries %}
+<h1>Choose a country to show the top 200 players in the country</h1>
+{% for country in countries %}
+<ul>
+    <li><a href="{{ url_for('player_ranking_data', country_code=country) }}">{{ country }}</a></li>
+</ul>
+{% endfor %}
+
+{% else %}
+<h1>No countries found</h1>
+{% endif %}
+
+</body>
+</html>
