+++ web_ui_service/templates/player_data.html
@@ -3,11 +3,11 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Brawlstars Player Data</title>
+    <title>Top 200 Brawlstars Player Data for {{ country_code }}</title>
     <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}">
 </head>
 <body>
-    <h1>Brawlstars Player Data</h1>
+    <h1>Brawlstars Player Data for {{ country_code }}</h1>
 
     {% if data %}
     <table>
